refactored envelopes
