+++ src/api/envelope/model.ts
@@ -1,4 +1,3 @@
-import { IConstructionDetail } from "src/models/construction";
 import { IConstructionSchema } from "../construction/models";
 
 export interface IEnvelopeSchema {
@@ -13,13 +12,4 @@ export interface IEnvelopeSchema {
   window: IConstructionSchema;
 }
 
-export interface IEnvelope_post {
-  name: string;
-  description: string;
-  exteriorWall: IConstructionDetail;
-  floorCeiling: IConstructionDetail;
-  groundFloor: IConstructionDetail;
-  interiorWall: IConstructionDetail;
-  roof: IConstructionDetail;
-  window: IConstructionDetail;
-}
+export type TEnvelopePayload = Omit<IEnvelopeSchema, "id">;
